server/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
diffusers>=0.27.2
transformers>=4.40.0
torch>=2.0.0
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.26.0
bitsandbytes>=0.43.0
accelerate>=0.30.0